==> scripts/02_03_trend_drivers_sen.Rmd <==
---
title: "trend drivers"
output: pdf_document
date: "2025-03-12"
author: "Junna Wang"
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```


```{r raw data location}
rm(list=ls())
# change this location where the raw data will be saved accordingly
# dir_rawdata <- '/Users/jw2946/Documents/stability_project/SiteData/'
dir_rawdata <- '/Volumes/MaloneLab/Research/Stability_Project/SiteData'
#
library(librarian)
shelf(tidyverse, ggpubr, MuMIn, car, fBasics, olsrr, lme4, Hmisc, corrplot, trend)
#
```


```{r get nitrogen deposition}
# data is from 
# Zhu (2025): Changing patterns of global nitrogen deposition driven by socio-economic development
Nfiles <- list.files('data/Global_N_deposition_grid_dataset_2008_2020', pattern='^mean_totN_', full.names = T)

TNdep2008_2020 <- terra::rast(Nfiles)

# transform to WGS 84 (longitude, latitude)
data_ll <- suppressWarnings(terra::project(TNdep2008_2020, "EPSG:4326"))

# get total N values from these rasters at our study sites
sites_long <- read.csv('data/long_term_ec_sites.csv')

xy <- data.frame(x=sites_long$Long, y=sites_long$Lat)
TNvalues <- terra::extract(data_ll, xy)

# deal with missing data
for (i in which(is.na(TNvalues$mean_totN_2008_hm))) {
  xy[i,] <- xy[i,] - 0.1482702
  TNvalues[i,2:ncol(TNvalues)] <- terra::extract(data_ll, xy[i,])[2:ncol(TNvalues)]
}

TNvalues$Nmean <- rowMeans(TNvalues)
TNvalues$Ntrend_lm <- 0
TNvalues$Np <- 0

for (i in 1:nrow(TNvalues)) {
  data <- data.frame(x=2008:2020, y=as.numeric(TNvalues[i,2:14]))
  TNvalues[i, 16:17] <- summary(lm(data=data, scale(y) ~ scale(x, scale=FALSE)))$coefficients[2, c(1,4)]
}

TNvalues$site_ID <- sites_long$site_ID

```


```{r get forest ages}

forestAge <- read.csv('data/forest_ages.csv')

tmp <- separate(forestAge, Years, into = c("start_year", "end_year"), sep = "-")

# process data in the first 126 row first.
# these 126 data is from the paper: Quantifying the effect of forest age in annual net forest carbon balance
tmp$end_year[is.na(tmp$end_year[1:126])] <- tmp$start_year[is.na(tmp$end_year[1:126])]

forestAge$inyear[1:126] <- floor((as.numeric(tmp$end_year[1:126]) + as.numeric(tmp$start_year[1:126])) / 2)

forestAge$Age_standard <- forestAge$Age - forestAge$inyear         # age compare to 0000 year

```

```{r get canopy height}
# plant height information
canopyHeightLang <- read.csv('data/CanopyHeight_Lang.csv')
canopyHeightChu <- read.csv('data/AMF-HEIGHTC-BADM_AND_PI_chu.csv')
#
sub_time_series <- read.csv('data/sub_time_series_sen.csv')
#
canopyHeight <- sub_time_series[, c("site_ID", "year_start", "year_end")]
canopyHeight$year_mean <- (canopyHeight$year_start + canopyHeight$year_end) / 2
canopyHeightChu <- canopyHeightChu[canopyHeightChu$SITE_ID %in% unique(canopyHeight$site_ID), ]
#
canopyHeightChu <- left_join(canopyHeightChu, canopyHeight[,c("site_ID", "year_mean")], by=c("SITE_ID"="site_ID"))
canopyHeightChu <- canopyHeightChu %>% mutate(year_dif=abs(YEAR-year_mean)) %>% group_by(SITE_ID, year_mean) %>% summarise(Height=HEIGHTC[which.min(year_dif)])
#
canopyHeight <- canopyHeight %>% left_join(canopyHeightChu, by=c("site_ID"="SITE_ID", "year_mean"))
# canopy height of CL-SDF is ~25m
canopyHeight$Height[canopyHeight$site_ID=="CL-SDF"] <- 25.0	
#
canopyHeight <- canopyHeight %>% left_join(canopyHeightLang[, c("site_ID", "canopyH_Lang")], by="site_ID")

canopyHeight$Height[is.na(canopyHeight$Height)] <- canopyHeight$canopyH_Lang[is.na(canopyHeight$Height)]

```


```{r doy of southern hemisphere}
# this function calculates day of year for the southern hemisphere using July 1st as doy = 1
southern_hemisphere_doy <- function(dates) {
  year_start <-
    ymd(paste(year(dates), "07-01", sep = "-"))  # July 1 of the same year
  doy <- as.numeric(dates - year_start + 1)  # Compute DOY
  
  # Adjust for dates before July 1 (shift to previous year's DOY count)
  doy[doy < 1] <-
    doy[doy < 1] + 365 + leap_year(year(dates[doy < 1]) - 1)
  
  return(doy)
}
#
```


```{r identify outlier years}
# This function calculates outlier years based on percentiles. 
OutlierYearsTA <- function(years, values) {
  # rule 1: no occasional outliers
  Q1 <- quantile(values, 0.25, na.rm=T)
  Q3 <- quantile(values, 0.75, na.rm=T)
  # IQR should be at least 2 C
  IQR <- max(Q3 - Q1, 1.8)
  outlier <- !between(values, Q1 - 1.5 * IQR, Q3 + 1.5 * IQR)
  if (length(outlier) > 0) {
    OutlierYearsTA <- years[outlier]    
  }
  # rule 2: no jumping outliers
  Q2  <- quantile(values, 0.5, na.rm=T)
  IQR <- max((Q2 - Q1) * 2, 1.8)
  outlier <- !between(values, Q1 - 1.5 * IQR, Q2 + Q2 - Q1 + 1.5 * IQR)
  if (length(outlier) > 0) {
    OutlierYearsTA <- unique(c(OutlierYearsTA, years[outlier]))
  }
  return(OutlierYearsTA[!is.na(OutlierYearsTA)])
}
#
# the function to get LE outlier
OutlierYearsLE <- function(years, values) {
  # rule 1: no occasional outliers
  Q1 <- quantile(values, 0.25, na.rm=T)
  Q3 <- quantile(values, 0.75, na.rm=T)
  IQR <- Q3 - Q1
  outlier <- !between(values, Q1 - 2.5 * IQR, Q3 + 2.5 * IQR)
  if (length(outlier) > 0) {
    OutlierYearsLE <- years[outlier]    
  }
  return(OutlierYearsLE[!is.na(OutlierYearsLE)])
}

```


```{r a function to merge ERA and FLUX DATA}
# merge ERA and FLUX data
merge_ERA_FLUX <- function(years2update, col_year, col_flux, col_era) {
  df <- data.frame(year=col_year, flux=col_flux, era=col_era)
  id2update <- df$year %in% years2update
  #
  df.sub <- df[!id2update, ]
  mod <- lm(data=df.sub, flux ~ era)
  #
  df$flux[id2update] <- predict(mod, newdata=df[id2update, ])
  return(df$flux)
}

```


```{r a function to calculate sens slope}
cal_sens_slope <- function(x) {
  x <- x[!is.na(x)]
  x <- scale(x)
  return(as.numeric(sens.slope(x)[c(1,3)]))
}

```


```{r trend of environmental conditions}
# it takes 5 mins. 
#### attention: run the three chunks/function above this chunk first ####

# read sub time series
sub_time_series <- read.csv('data/sub_time_series_sen.csv')
#
data_grow_season <- read.csv('data/growing_season_yearly.csv')
#
sites_long <- read.csv('data/long_term_ec_sites.csv')
#
rm_years   <- read.csv('data/years_to_remove.csv')
#
rm_years_co2 <- read.csv('data/bad_co2_site_year.csv')
#
# add data source to sub_time_series
if (!"source" %in% colnames(sub_time_series)) {
  sub_time_series <- merge(sub_time_series, sites_long[, c('site_ID', "source", "Lat")], by="site_ID")
}

# mutate many columns to sub_time_series
sub_time_series <- sub_time_series %>% mutate(TA_t=NA, TA_p=NA, TA_gs_t=NA, TA_gs_p=NA, TA_ngs_t=NA, TA_ngs_p=NA,
                   P_t=NA, P_p=NA, P_gs_t=NA, P_gs_p=NA, P_ngs_t=NA, P_ngs_p=NA,
                   VPD_t=NA, VPD_p=NA, VPD_gs_t=NA, VPD_gs_p=NA, VPD_ngs_t=NA, VPD_ngs_p=NA,
                   SW_t=NA, SW_p=NA, SW_gs_t=NA, SW_gs_p=NA, SW_ngs_t=NA, SW_ngs_p=NA,
                   LE_t=NA, LE_p=NA, LE_gs_t=NA, LE_gs_p=NA, LE_ngs_t=NA, LE_ngs_p=NA,
                   CO2_t=NA, CO2_p=NA, Lgs_t=NA, Lgs_p=NA, Lgs_piece_t=NA, Lgs_piece_p=NA,
                   MAT=NA, MAP=NA, Height=NA)

data_trend_driver <- data.frame()

# get the average start and end growing season for each site
grow_season_site <- data_grow_season %>%
  group_by(site_ID) %>% 
  summarise(sos=floor((mean(sos_DER) + mean(sos_TRS)) / 2), 
            eos=ceiling((mean(eos_DER) + mean(eos_TRS)) / 2))
########################Important information here########################
# remember that one site CL-SDF is in the southern hemisphere; 
# year long growing season: 'BR-Sa1', 'US-Esm', 'US-LL1', 'US-LL2', 'US-LL3', 'US-SP3', 'GF-Guy';
# for these sites we assume growing and non-growing season have the same trend
########################################################################## 
# prepare for dataset
files.unzip.Ameri <- list.files(file.path(dir_rawdata, "Ameriflux_FLUXNET", "unzip"), full.names = T)
files.unzip.Europ2020 <- list.files(file.path(dir_rawdata, "FLUXNET2020", "unzip"), full.names = T)
files.unzip.Europ2023 <- list.files(file.path(dir_rawdata, "ICOS_FLUXNET", "unzip_connect2020"), full.names = T)
files.Ameri.ReddyProcGapFill <- list.files(file.path(dir_rawdata, "Ameriflux_BASE", "ReddyProcGapFill"), full.names=T)

# total number of site-years removed
TA_site_years_removed <- 0

# set na options for linear regression
options(na.action = na.omit)

## calculate trends of environmental drivers
for (i in 1:nrow(sub_time_series)) {
  print(i)
  site_ID <- sub_time_series$site_ID[i]
  print(site_ID)
  ##
  if (sub_time_series$source[i] == "AmeriFlux_FLUXNET") {
    a_YY <- read.csv(files.unzip.Ameri[grepl(pattern=paste0(site_ID, "_FLUXNET_FULLSET_YY"), files.unzip.Ameri)])
    a_DD <- read.csv(files.unzip.Ameri[grepl(pattern=paste0(site_ID, "_FLUXNET_FULLSET_DD"), files.unzip.Ameri)])
  } else if (sub_time_series$source[i] == "EuropFlux_FLUXNET2020") {
    a_YY <- read.csv(files.unzip.Europ2020[grepl(pattern=paste0(site_ID, "_FLUXNET2015_FULLSET_YY"), files.unzip.Europ2020)])
    a_DD <- read.csv(files.unzip.Europ2020[grepl(pattern=paste0(site_ID, "_FLUXNET2015_FULLSET_DD"), files.unzip.Europ2020)])
  } else if (sub_time_series$source[i] == "EuropFlux_FLUXNET2023") {
    a_YY <- read.csv(files.unzip.Europ2023[grepl(pattern=paste0(site_ID, "_FLUXNET_YY"), files.unzip.Europ2023)])
    a_DD <- read.csv(files.unzip.Europ2023[grepl(pattern=paste0(site_ID, "_FLUXNET_DD"), files.unzip.Europ2023)])
  } else if (sub_time_series$source[i] == "AmeriFlux_BASE") {
    a_YY <- read.csv(files.Ameri.ReddyProcGapFill[grepl(pattern=paste0(site_ID, "_EDDYPROC_FULLSET_YY"), files.Ameri.ReddyProcGapFill)])
    a_DD <- read.csv(files.Ameri.ReddyProcGapFill[grepl(pattern=paste0(site_ID, "_EDDYPROC_FULLSET_DD"), files.Ameri.ReddyProcGapFill)])
  }
  a_YY[a_YY==-9999] <- NA
  a_DD[a_DD==-9999] <- NA
  #
  a_YY <- a_YY %>% filter(!TIMESTAMP %in% rm_years$years[rm_years$site_ID==site_ID]) %>%
    filter(TIMESTAMP >= sub_time_series$year_start[i] & TIMESTAMP <= sub_time_series$year_end[i])
  #
  a_DD$TIMESTAMP <- as.Date(as.character(a_DD$TIMESTAMP), "%Y%m%d")
  a_DD <- a_DD %>% mutate(year=year(a_DD$TIMESTAMP), doy=yday(a_DD$TIMESTAMP))
  if (sub_time_series$Lat[i] < 0.0) {
    a_DD$doy <- southern_hemisphere_doy(a_DD$TIMESTAMP)
  }
  #
  # remove bad years and select the correct years
  a_DD <- a_DD %>% filter(!year %in% rm_years$years[rm_years$site_ID==site_ID]) %>% 
    filter(year >= sub_time_series$year_start[i] & year <= sub_time_series$year_end[i])
  # print(plot(a_DD$CO2_F_MDS, main=site_ID))
  #
  # deal with missing variables
  if (sum(!is.na(a_YY$VPD_F_MDS)) < 5)  {
    a_YY$VPD_F_MDS <- a_YY$VPD_F
    a_DD$VPD_F_MDS <- a_DD$VPD_F
  }
  if (sum(!is.na(a_YY$SW_IN_F_MDS)) < 5)  {
    a_YY$SW_IN_F_MDS <- a_YY$SW_IN_F
    a_DD$SW_IN_F_MDS <- a_DD$SW_IN_F
  }
  
  # remove bad CO2 data
  if (! 'CO2_F_MDS' %in% names(a_YY) ) {
    a_YY$CO2_F_MDS <- NA
  } else {
    if (site_ID %in% rm_years_co2$site_ID) {
      rm_years_co2_1site = rm_years_co2[rm_years_co2$site_ID==site_ID, 2:ncol(rm_years_co2)]
      rm_years_co2_1site = rm_years_co2_1site[!is.na(rm_years_co2_1site)]
      a_YY$CO2_F_MDS[a_YY$TIMESTAMP %in% rm_years_co2_1site] <- NA
    }
    if ( sum(!is.na(a_YY$CO2_F_MDS)) < 5 ) {
      print(paste0('not enought CO2 data-', site_ID))
    }
  }
  # # temporally output the CO2 time series
  if (!all(is.na(a_YY$CO2_F_MDS))) {
    plot(a_DD$TIMESTAMP, a_DD$CO2_F_MDS, main =site_ID)
  }
  
  #####This is a big section to remove outliers in environmental conditions#####
  # replace TA and VPD outliers with ERA5 data
  outlier_years_TA <- OutlierYearsTA(a_YY$TIMESTAMP, a_YY$TA_F_MDS)
  if (length(outlier_years_TA) > 0) {
    TA_site_years_removed <- TA_site_years_removed + length(outlier_years_TA)
    print(paste0('Site ', site_ID, ' has TA outliers in years ', outlier_years_TA))
    #
    if (!"TA_ERA" %in% names(a_DD)) {
        a_DD$TA_F_MDS[a_DD$year %in% outlier_years_TA] <- NA
        a_DD$VPD_F_MDS[a_DD$year %in% outlier_years_TA] <- NA
        a_YY$TA_F_MDS[a_YY$TIMESTAMP %in% outlier_years_TA] <- NA
        a_YY$VPD_F_MDS[a_YY$TIMESTAMP %in% outlier_years_TA] <- NA
        print('no_ERA5')
    } else {
      if (length(outlier_years_TA) <= 2) {
        # replace with ERA5 data after doing linear regression
        a_DD$TA_F_MDS <- merge_ERA_FLUX(outlier_years_TA, a_DD$year, a_DD$TA_F_MDS, a_DD$TA_ERA)
        a_DD$VPD_F_MDS <- merge_ERA_FLUX(outlier_years_TA, a_DD$year, a_DD$VPD_F_MDS, a_DD$VPD_ERA)
        a_YY$TA_F_MDS <- merge_ERA_FLUX(outlier_years_TA, a_YY$TIMESTAMP, a_YY$TA_F_MDS, a_YY$TA_ERA)
        a_YY$VPD_F_MDS <- merge_ERA_FLUX(outlier_years_TA, a_YY$TIMESTAMP, a_YY$VPD_F_MDS, a_YY$VPD_ERA)
      } else if (length(outlier_years_TA) > 2) {
        # directly use ERA5 data
        a_DD$TA_F_MDS <- a_DD$TA_ERA
        a_DD$VPD_F_MDS <- a_DD$VPD_ERA
        a_YY$TA_F_MDS <- a_YY$TA_ERA
        a_YY$VPD_F_MDS <- a_YY$VPD_ERA
      }
    }    
  }

  # remove extreme LE outliers using the simplest method
  outlier_years_LE <- OutlierYearsLE(a_YY$TIMESTAMP, a_YY$LE_F_MDS)
  
  # remove years with bad data quality based QC flag
  if ('LE_F_MDS_QC' %in% names(a_YY)) {
    lackdata_years <- a_YY$TIMESTAMP[a_YY$LE_F_MDS_QC < 0.5 & !is.na(a_YY$LE_F_MDS_QC)]
    if (length(lackdata_years) > 0) {
      outlier_years_LE <- unique(c(outlier_years_LE, lackdata_years))
    }
  }
  
  # set NA to these years' LE
  if (length(outlier_years_LE) > 0) {
    a_DD$LE_F_MDS[a_DD$year %in% outlier_years_LE] <- NA
    a_YY$LE_F_MDS[a_YY$TIMESTAMP %in% outlier_years_LE] <- NA
    print(paste0('Site ', site_ID, ' has LE outliers in years ', outlier_years_LE))
  }
  #
  # #############start plot figures for data quality check##########
  # plot(a_DD$TIMESTAMP, a_DD$TA_F_MDS, main=site_ID)
  # plot(a_YY$TIMESTAMP, a_YY$TA_F_MDS, main=site_ID)
  # #
  # plot(a_DD$TIMESTAMP, a_DD$VPD_F_MDS, main=site_ID)
  # plot(a_YY$TIMESTAMP, a_YY$VPD_F_MDS, main=site_ID)  
  # #
  # plot(a_DD$TIMESTAMP, a_DD$SW_IN_F_MDS, main=site_ID)
  # plot(a_YY$TIMESTAMP, a_YY$SW_IN_F_MDS, main=site_ID)
  # #
  # plot(a_DD$TIMESTAMP, a_DD$LE_F_MDS, main=site_ID)
  # plot(a_YY$TIMESTAMP, a_YY$LE_F_MDS, main=site_ID)
  # #
  # plot(a_DD$TIMESTAMP, a_DD$P_ERA, main=site_ID)
  # plot(a_YY$TIMESTAMP, a_YY$P_ERA, main=site_ID)
  # #
  # plot(a_DD$TIMESTAMP, a_DD$CO2_F_MDS, main=site_ID)
  # plot(a_YY$TIMESTAMP, a_YY$CO2_F_MDS, main=site_ID)
  # #############end plot figures for data quality check##########
  
  ## we differentiate TA, P, VPD, SW_IN, LE for growing and non-growing seasons
  a_YY$year_center <- as.numeric(scale(a_YY$TIMESTAMP, scale=FALSE))
  if (site_ID %in% c('BR-Sa1', 'US-Esm', "US-Elm", 'US-LL1', 'US-LL2', 'US-LL3', 'US-SP3', 'GF-Guy')) {
    a_YY_gw <- a_YY
    a_YY_ngw <- a_YY
  } else {
    # start and end of growing season
    sos <- grow_season_site$sos[grow_season_site$site_ID == site_ID]
    eos <- grow_season_site$eos[grow_season_site$site_ID == site_ID]
    a_YY_gw <- a_DD %>% filter(doy >= sos & doy <= eos) %>% group_by(year) %>% 
          summarise(TA_F_MDS=mean(TA_F_MDS, na.rm=T), P_ERA=sum(P_ERA, na.rm=T), 
                    VPD_F_MDS=mean(VPD_F_MDS, na.rm=T), SW_IN_F_MDS=mean(SW_IN_F_MDS, na.rm=T), 
                    LE_F_MDS=mean(LE_F_MDS, na.rm=T))
    a_YY_ngw <- a_DD %>% filter(doy < sos | doy > eos) %>% group_by(year) %>% 
          summarise(TA_F_MDS=mean(TA_F_MDS, na.rm=T), P_ERA=sum(P_ERA, na.rm=T), 
                    VPD_F_MDS=mean(VPD_F_MDS, na.rm=T), SW_IN_F_MDS=mean(SW_IN_F_MDS, na.rm=T), 
                    LE_F_MDS=mean(LE_F_MDS, na.rm=T))
    
    # deal with missed P_ERA data in some years
    missed_PERA_years <- a_YY$TIMESTAMP[is.na(a_YY$P_ERA)]
    a_YY_gw$P_ERA[a_YY_gw$year %in% missed_PERA_years] <- NA
    a_YY_ngw$P_ERA[a_YY_ngw$year %in% missed_PERA_years] <- NA
    #
    a_YY_gw$year_center  <- as.numeric(scale(a_YY_gw$year, scale=FALSE))
    a_YY_ngw$year_center <- as.numeric(scale(a_YY_ngw$year, scale=FALSE))
  }
  #
  # Calculate trends
  sub_time_series[i, c("TA_t", "TA_p")] <- cal_sens_slope(a_YY$TA_F_MDS)
    # summary(lm(data=a_YY, scale(TA_F_MDS) ~ year_center))$coefficients[2, c(1,4)]
  sub_time_series[i, c("TA_gs_t", "TA_gs_p")] <- cal_sens_slope(a_YY_gw$TA_F_MDS)
    # summary(lm(data=a_YY_gw, scale(TA_F_MDS) ~ year_center))$coefficients[2, c(1,4)]
  sub_time_series[i, c("TA_ngs_t", "TA_ngs_p")] <- cal_sens_slope(a_YY_ngw$TA_F_MDS)
    # summary(lm(data=a_YY_ngw, scale(TA_F_MDS) ~ year_center))$coefficients[2, c(1,4)]
  #
  sub_time_series[i, c("P_t", "P_p")] <- cal_sens_slope(a_YY$P_ERA)
    # summary(lm(data=a_YY, scale(P_ERA) ~ year_center))$coefficients[2, c(1,4)]
  sub_time_series[i, c("P_gs_t", "P_gs_p")] <- cal_sens_slope(a_YY_gw$P_ERA)
    # summary(lm(data=a_YY_gw, scale(P_ERA) ~ year_center))$coefficients[2, c(1,4)]
  sub_time_series[i, c("P_ngs_t", "P_ngs_p")] <- cal_sens_slope(a_YY_ngw$P_ERA)
    # summary(lm(data=a_YY_ngw, scale(P_ERA) ~ year_center))$coefficients[2, c(1,4)]
  #
  sub_time_series[i, c("VPD_t", "VPD_p")] <- cal_sens_slope(a_YY$VPD_F_MDS)
    # summary(lm(data=a_YY, scale(VPD_F_MDS) ~ year_center))$coefficients[2, c(1,4)]
  sub_time_series[i, c("VPD_gs_t", "VPD_gs_p")] <- cal_sens_slope(a_YY_gw$VPD_F_MDS)
    # summary(lm(data=a_YY_gw, scale(VPD_F_MDS) ~ year_center))$coefficients[2, c(1,4)]
  sub_time_series[i, c("VPD_ngs_t", "VPD_ngs_p")] <- cal_sens_slope(a_YY_ngw$VPD_F_MDS)
    # summary(lm(data=a_YY_ngw, scale(VPD_F_MDS) ~ year_center))$coefficients[2, c(1,4)]
  #
  sub_time_series[i, c("SW_t", "SW_p")] <- cal_sens_slope(a_YY$SW_IN_F_MDS)
    # summary(lm(data=a_YY, scale(SW_IN_F_MDS) ~ year_center))$coefficients[2, c(1,4)]
  sub_time_series[i, c("SW_gs_t", "SW_gs_p")] <- cal_sens_slope(a_YY_gw$SW_IN_F_MDS)
    # summary(lm(data=a_YY_gw, scale(SW_IN_F_MDS) ~ year_center))$coefficients[2, c(1,4)]
  sub_time_series[i, c("SW_ngs_t", "SW_ngs_p")] <- cal_sens_slope(a_YY_ngw$SW_IN_F_MDS)
    # summary(lm(data=a_YY_ngw, scale(SW_IN_F_MDS) ~ year_center))$coefficients[2, c(1,4)]
  #
  sub_time_series[i, c("LE_t", "LE_p")] <- cal_sens_slope(a_YY$LE_F_MDS)
    # summary(lm(data=a_YY, scale(LE_F_MDS) ~ year_center))$coefficients[2, c(1,4)]
  sub_time_series[i, c("LE_gs_t", "LE_gs_p")] <- cal_sens_slope(a_YY_gw$LE_F_MDS)
    # summary(lm(data=a_YY_gw, scale(LE_F_MDS) ~ year_center))$coefficients[2, c(1,4)]
  sub_time_series[i, c("LE_ngs_t", "LE_ngs_p")] <- cal_sens_slope(a_YY_ngw$LE_F_MDS)
    # summary(lm(data=a_YY_ngw, scale(LE_F_MDS) ~ year_center))$coefficients[2, c(1,4)]
  #
  # left join a_YY, a_YY_gw, and a_YY_ngw and put data into a data frame
  a_YY_gw_rename <- a_YY_gw
  a_YY_ngw_rename <- a_YY_ngw
  if ("year" %in% names(a_YY_gw_rename)) {
    a_YY_gw_rename <- a_YY_gw_rename %>% rename('TIMESTAMP' = "year")
  }
  if ("year" %in% names(a_YY_ngw_rename)) {
    a_YY_ngw_rename <- a_YY_ngw_rename %>% rename('TIMESTAMP' = "year")
  }
  #
  varnames_select <- c("TIMESTAMP", "TA_F_MDS", "P_ERA", "VPD_F_MDS", "SW_IN_F_MDS", "LE_F_MDS")
  a_YY_gw_rename <- a_YY_gw_rename[, varnames_select]
  names(a_YY_gw_rename)[2:6] <- paste0(varnames_select[2:6], "_gw")
  a_YY_ngw_rename <- a_YY_ngw_rename[, varnames_select]
  names(a_YY_ngw_rename)[2:6] <- paste0(varnames_select[2:6], "_ngw")
  #
  merged_df <- merge(a_YY[, c("NEE_VUT_REF", varnames_select)], a_YY_gw_rename, by="TIMESTAMP")
  merged_df <- merge(merged_df, a_YY_ngw_rename, by="TIMESTAMP")
  merged_df$site_ID <- site_ID
  merged_df$isubts <- i
  data_trend_driver <- rbind(data_trend_driver, merged_df)
  #
  ## we do not differentiate CO2, Lgs for growing and non-growing seasons
  if (sum(!is.na(a_YY$CO2_F_MDS)) > 2) {
#    sub_time_series[i, c("CO2_t", "CO2_p")] <- cal_sens_slope(a_YY$CO2_F_MDS)
    # I should not standarize CO2, because I want to know how per ppm CO2 increase affect NEE trend. 
    sub_time_series[i, c("CO2_t", "CO2_p")] <- summary(lm(data=a_YY, CO2_F_MDS ~ year_center))$coefficients[2, c(1,4)]
    # deal with abnormal data; the standard deviation is 0.9 ppm/year; mean = 2.36 ppm/year
    if (sub_time_series$CO2_t[i] > 4.3 | sub_time_series$CO2_t[i] < 0.5) {
      sub_time_series[i, c("CO2_t", "CO2_p")] <- NA
    }
  }
  if (site_ID %in% c('BR-Sa1', 'US-Esm', "US-Elm", 'US-LL1', 'US-LL2', 'US-LL3', 'US-SP3', 'GF-Guy')) {
    sub_time_series[i, c("Lgs_t", "Lgs_p")] <- 0
    sub_time_series[i, c("Lgs_piece_t", "Lgs_piece_p")] <- 0
  } else {
    subdata_grow_season <- data_grow_season[data_grow_season$site_ID==site_ID, ] %>% filter(year >= sub_time_series$year_start[i] & year <= sub_time_series$year_end[i])
    if (nrow(subdata_grow_season) >= 5) {
      sub_time_series[i, c("Lgs_t", "Lgs_p")] <- cal_sens_slope(subdata_grow_season$length_gs)
        # summary(lm(data=subdata_grow_season, scale(length_gs) ~ scale(year, scale=FALSE)))$coefficients[2, c(1,4)]
      sub_time_series[i, c("Lgs_piece_t", "Lgs_piece_p")] <- cal_sens_slope(subdata_grow_season$length_gs_piece)
        # summary(lm(data=subdata_grow_season, scale(length_gs_piece) ~ scale(year, scale=FALSE)))$coefficients[2, c(1,4)]
    }
    
    # ###########################start plot figures for data quality check##################
    # plot(subdata_grow_season$year, subdata_grow_season$length_gs, main=site_ID)
    # plot(subdata_grow_season$year, subdata_grow_season$length_gs_piece, main=site_ID)
    # ###########################end plot figures for data quality check####################
    
  }
  # Baseline conditions: MAT, MAP; use our TNvalues if we have >10 year data
  if (is.na(sites_long$MAT[sites_long$site_ID==site_ID])) {
    sub_time_series[i, "MAT"] <- mean(a_YY$TA_F_MDS, na.rm=T)
    sub_time_series[i, "MAP"] <- mean(a_YY$P_ERA, na.rm=T)    
  } else {
    sub_time_series[i, "MAT"] <- sites_long$MAT[sites_long$site_ID==site_ID]
    sub_time_series[i, "MAP"] <- sites_long$MAP[sites_long$site_ID==site_ID]
  }
  # plant height
  sub_time_series[i, "Height"] <- canopyHeight$Height[canopyHeight$site_ID==site_ID & abs(canopyHeight$year_start-sub_time_series$year_start[i]) < 0.1]
}
#
# add forest age
if (!"Age" %in% colnames(sub_time_series)) {
  sub_time_series <- left_join(sub_time_series, forestAge[, c('Site', "Age_standard")], by=c("site_ID"="Site"))
  sub_time_series$Age <- sub_time_series$Age_standard + (sub_time_series$year_start + sub_time_series$year_end) / 2.0
  sub_time_series$Age_standard <- NULL
}
#
# add TN deposition data
if (!"Nmean" %in% colnames(sub_time_series)) {
  sub_time_series <- left_join(sub_time_series, TNvalues[, c('site_ID', "Nmean", "Ntrend_lm", "Np")], by="site_ID")
}

if (!"IGBP" %in% colnames(sub_time_series)) {
  sub_time_series <- merge(sub_time_series, sites_long[, c('site_ID', "IGBP")], by="site_ID")
}
#
print(paste0('total number of site years with TA outliers is removed: ', TA_site_years_removed))

# remove poor-quality CO2 data
sub_time_series$CO2_t[sub_time_series$CO2_t < 0.05] <- NA

write.csv(sub_time_series, "data/sub_time_series_potential_drivers_sen.csv", row.names = F)

write.csv(data_trend_driver, "data/data_trend_driver_sen.csv", row.names = F)

```


```{r any relationships between CO2 and NEE trend}
if (! "category" %in% names(sub_time_series)) {
  sub_time_series <- sub_time_series %>% 
    left_join(sites_long[, c("site_ID", "category")], by="site_ID")
}

sub_time_series %>% 
  filter(IGBP %in% c('ENF', 'EBF') & category == "cold") %>% 
  ggplot(aes(x=CO2_t, y=NEEtrend_lm)) +
  geom_point() +
  geom_smooth(method='lm') +
  stat_cor()

sub_time_series %>% 
  filter(IGBP %in% c('DBF', 'MF') & category == "cold") %>% 
  ggplot(aes(x=CO2_t, y=NEEtrend_lm, color=IGBP)) +
  geom_point() +
  geom_smooth(method='lm') +
  stat_cor()

sub_time_series %>% 
  filter(!IGBP %in% c('ENF', 'EBF', 'DBF', 'MF') & category == "cold") %>% 
  ggplot(aes(x=CO2_t, y=NEEtrend_lm)) +
  geom_point() +
  geom_smooth(method='lm') +
  stat_cor()

sub_time_series %>% 
  filter(category == "wet") %>% 
  ggplot(aes(x=CO2_t, y=NEEtrend_lm)) +
  geom_point() +
  geom_smooth(method='lm') +
  stat_cor()

sub_time_series %>% 
  filter(category == "dry") %>% 
  ggplot(aes(x=CO2_t, y=NEEtrend_lm)) +
  geom_point() +
  geom_smooth(method='lm') +
  stat_cor()

###I may detect fertilization effect only if I have many sites for each climate. it is very weak in this data. 

###How about correlation with GPP; The relationship is converse;
sub_time_series %>% 
  filter(IGBP %in% c('ENF', 'EBF') & category == "cold") %>% 
  ggplot(aes(x=CO2_t, y=GPPtrend_lm)) +
  geom_point() +
  geom_smooth(method='lm') +
  stat_cor()

sub_time_series %>% 
  filter(!IGBP %in% c('ENF', 'EBF', 'DBF', 'MF') & category == "cold") %>% 
  ggplot(aes(x=CO2_t, y=GPPtrend_lm)) +
  geom_point() +
  geom_smooth(method='lm') +
  stat_cor()


### but climate's effect is there; local CO2 concentration has effects on TA and SW trend
sub_time_series %>% 
  filter(IGBP %in% c('ENF', 'EBF') & category == "cold") %>% 
  ggplot(aes(x=CO2_t, y=TA_gs_t)) +
  geom_point() +
  geom_smooth(method='lm') +
  stat_cor()

sub_time_series %>% 
  filter(IGBP %in% c('DBF', 'MF') & category == "cold") %>% 
  ggplot(aes(x=CO2_t, y=TA_gs_t, col=IGBP)) +
  geom_point() +
  geom_smooth(method='lm') +
  stat_cor()

# there is global spatial heterogeneity there; why they are different?
sub_time_series %>% 
  filter(!IGBP %in% c('ENF', 'EBF', 'DBF', 'MF') & category == "cold") %>% 
  ggplot(aes(x=CO2_t, y=TA_gs_t, col=IGBP)) +
  geom_point() +
  geom_smooth(method='lm') +
  stat_cor()

sub_time_series %>% 
  filter(category == "wet") %>% 
  ggplot(aes(x=CO2_t, y=TA_gs_t)) +
  geom_point() +
  geom_smooth(method='lm') +
  stat_cor()

# this is related to vegetation types! no effects. 
sub_time_series %>% 
  filter(category == "dry") %>% 
  ggplot(aes(x=CO2_t, y=TA_gs_t, col=IGBP)) +
  geom_point() +
  geom_smooth(method='lm') +
  stat_cor()


```


```{r drivers of trend in cold climate, echo=TRUE, fig.height=8, fig.width=8}
#
data_cold <- sub_time_series %>% left_join(sites_long[, c("site_ID", "category")], by="site_ID") %>% 
  filter(category=='cold') 

data <- data_cold %>% mutate(Latabs = abs(Lat)) %>%
        dplyr::select(c("NEEtrend_lm", "GPPtrend_lm", "ERtrend_lm", 
                            "TA_t", "P_t", "VPD_t", "SW_t", "LE_t", 
                            "TA_gs_t", "P_gs_t", "VPD_gs_t", "SW_gs_t", "LE_gs_t", 
                            "TA_ngs_t", "P_ngs_t", "VPD_ngs_t", "SW_ngs_t", "LE_ngs_t", 
                            "Lgs_t", "Lgs_piece_t", "MAT", "MAP", "Height", "IGBP", "Nmean", "Ntrend_lm"))
# "CO2_t", 
data_use <- subset(data, select=-c(IGBP))
data_use <- na.omit(data_use)

rcorr_results <- rcorr(as.matrix(data_use))
cor_matrix <- rcorr_results$r
p_matrix <- rcorr_results$P

# Print correlation matrix
print(cor_matrix)

# Print p-value matrix
print(p_matrix)

#
M=cor(data_use)
testRes = cor.mtest(data_use, conf.level = 0.95)
#
plot <- corrplot(M, p.mat = testRes$p, method = 'circle', insig='blank',
         diag = FALSE)
plot

# What's wrong with the cold climate data? P and SW during the non-growing season matter?
options(na.action = na.fail)
#
data_use <- subset(data, select=-c(GPPtrend_lm, ERtrend_lm))
data_use <- na.omit(data_use)
#
model <- lm(NEEtrend_lm ~ ., data = data_use)
summary(model)
#
vif_values <- vif(model)
vif_values

# these pairs cannot coexist
sexpr <- expression(!( (Lgs_t && Lgs_piece_t) ||
                         (TA_t && TA_ngs_t) || (TA_t && TA_gs_t) ||
                         (P_t && P_ngs_t) || (P_t && P_gs_t) ||
                         (VPD_t && VPD_ngs_t) || (VPD_t && VPD_gs_t) ||
                         (SW_t && SW_ngs_t) || (SW_t && SW_gs_t) ||
                         (LE_t && LE_ngs_t) || (LE_t && LE_gs_t)))

system.time(dd <- dredge(model, subset = sexpr, rank = "AICc"))    # , cluster = cl
dd

model_best <- lm(NEEtrend_lm ~ P_ngs_t + SW_ngs_t + VPD_ngs_t + IGBP, data = data_use)
summary(model_best)
# this is the worse model

# Differentiate by vegetation types; None of them significant
data_use %>% # filter(IGBP=='OSH') %>%
  ggplot(aes(x=Lgs_t, y=NEEtrend_lm)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')
  

```

```{r the influence of forest age}
# Get all the forest sites with forest age, see if age matters. 
data <- sub_time_series %>% left_join(sites_long[, c("site_ID", "category")], by="site_ID") %>%  filter(IGBP %in% c('ENF', 'DBF', 'MF', 'EBF')) %>%
mutate(Latabs = abs(Lat)) %>%
        dplyr::select(c("NEEtrend_lm", "GPPtrend_lm", "ERtrend_lm", "category", "NEE", "GPP", "ER", 
                        "Lgs_t", "Lgs_piece_t", "MAT", "MAP", "Height", "Latabs", "IGBP", "Nmean", "Ntrend_lm", "Age"))


# Get all the forest sites with forest age in cold climate, see if age matters. 
data %>% filter(IGBP %in% c('ENF', 'DBF', 'MF', 'EBF')) %>%      # c('ENF', 'DBF', 'MF', 'EBF')
  # filter(category=='cold') %>% 
  ggplot(aes(x=Age, y=NEEtrend_lm, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('DBF', "MF")) %>%      # c('ENF', 'DBF', 'MF', 'EBF')
  # filter(category=='cold') %>% 
  ggplot(aes(x=Age, y=NEEtrend_lm, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('ENF', "EBF")) %>%      # c('ENF', 'DBF', 'MF', 'EBF')
  # filter(category=='cold') %>% 
  ggplot(aes(x=Age, y=NEEtrend_lm, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')


data %>% filter(IGBP %in% c('ENF', 'DBF', 'MF', 'EBF')) %>%      # c('ENF', 'DBF', 'MF', 'EBF')
  # filter(category=='cold') %>% 
  ggplot(aes(x=Age, y=NEE, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

#
data %>% filter(IGBP %in% c('ENF', 'DBF', 'MF')) %>%      # c('ENF', 'DBF', 'MF', 'EBF')
  # filter(category=='cold') %>% 
  ggplot(aes(x=Age, y=GPP, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

#
data %>% filter(IGBP %in% c('ENF', 'MF')) %>%      # c('ENF', 'DBF', 'MF', 'EBF')
  # filter(category=='cold') %>% 
  ggplot(aes(x=Age, y=ER, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

#
data %>% filter(IGBP %in% c('EBF')) %>%      # c('ENF', 'DBF', 'MF', 'EBF')
  # filter(category=='cold') %>% 
  ggplot(aes(x=Age, y=ER, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')


data %>% filter(IGBP %in% c('EBF')) %>%      # c('ENF', 'DBF', 'MF', 'EBF')
  # filter(category=='cold') %>% 
  ggplot(aes(x=Age, y=GPP, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')


# Mixed forests: Age has the opposite effects on NEE trends in mixed forests and DBF than Evergreen needle-leaf and Evergreen broaf forests. 
# Could this be that trees in evergreen forests live longer than trees in other forests?
# Important references: a 500-year evergreen conifer forests remain carbon sinks, but can switch between carbon sources and sinks. 
# reference: Wharton et al., 2012; Old-growth CO2 flux measurements reveal high sensitivity to climate anomalies. 
# I got forest age data from this paper Besnard et al., 2018: quntifying the effect of forest age in annual net forest carbon balance. 


# I really need mean NEE values. 
# for non-EBF ('MF', 'ENF', 'DBF'), GPP and ER of old-forest reduce with age. 


# for EBF, GPP and ER of old-forest increase with age. 


```


```{r the impact of canopy heights}
data <- sub_time_series %>% left_join(sites_long[, c("site_ID", "category")], by="site_ID") %>% 
mutate(Latabs = abs(Lat)) %>% filter(IGBP %in% c('ENF', 'DBF', 'MF', 'EBF')) %>% filter(!category %in% c('dry')) %>%
        dplyr::select(c("NEEtrend_lm", "GPPtrend_lm", "ERtrend_lm", "category", "NEE", "GPP", "ER", 
                        "Lgs_t", "Lgs_piece_t", "MAT", "MAP", "Height", "Latabs", "IGBP", "Nmean", "Ntrend_lm", "Age", "site_ID"))


# Get all the forest sites with forest age in cold climate, see if age matters. 
data %>% filter(IGBP %in% c('ENF', 'DBF', 'MF', 'EBF')) %>%      # c('ENF', 'DBF', 'MF', 'EBF')
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=NEEtrend_lm, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

#-----------------------DBF---------------------
# GPP, ER, and NEE is not sensitive to canopy height, but NEE trend is sensitive; higher canopy, more positive NEE trend.  
data %>% filter(IGBP %in% c('DBF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=NEEtrend_lm, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('DBF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=NEE, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('DBF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=GPP, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('DBF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=ER, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

#-----------------------ENF---------------------
# GPP, ER, NEE and NEE trend are all sensitive to canopy height; higher canopy, more negative NEE trend.  

data %>% filter(IGBP %in% c('ENF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=NEEtrend_lm, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('ENF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=NEE, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('ENF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=GPP, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('ENF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=ER, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

#-----------------------MF---------------------not enough data
data %>% filter(IGBP %in% c('MF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=NEEtrend_lm, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('MF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=NEE, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('MF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=GPP, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('MF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=ER, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

#-----------------------EBF---------------------not enough data
data %>% filter(IGBP %in% c('EBF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=NEEtrend_lm, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('EBF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=NEE, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('EBF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=GPP, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

data %>% filter(IGBP %in% c('EBF')) %>%      
  # filter(category=='cold') %>% 
  ggplot(aes(x=Height, y=ER, col=IGBP)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')

```


```{r trend driver for dry climate }
# identify trend drivers for dry climate
data_dry <- sub_time_series %>% left_join(sites_long[, c("site_ID", "category", "IGBP")], by="site_ID") %>% filter(category=='dry')

data <- data_dry %>% mutate(Latabs = abs(Lat)) %>%
        dplyr::select(c("NEEtrend_lm", "GPPtrend_lm", "ERtrend_lm",
                            "TA_gs_t", "P_gs_t", "VPD_gs_t", "SW_gs_t", "LE_gs_t", 
                            "TA_ngs_t", "P_ngs_t", "VPD_ngs_t", "SW_ngs_t", "LE_ngs_t", 
                            "Lgs_t", "Height"))  # , "Nmean" "Lgs_piece_t", "Height", "Nmean"

rcorr_results <- rcorr(as.matrix(data))
cor_matrix <- rcorr_results$r
p_matrix <- rcorr_results$P

# Print correlation matrix
print(cor_matrix)

# Print p-value matrix
print(p_matrix)

#
data_use <- na.omit(data)
#
M=cor(data_use)
testRes = cor.mtest(data_use, conf.level = 0.95)
#
plot <- corrplot(M, p.mat = testRes$p, method = 'circle', insig='blank',
         diag = FALSE)
plot
#
options(na.action = na.fail)
#
# TA variables, MAT, MAP, Height are NOT important, can only have 18 variables in. TA_t, TA_gs_t, TA_ngs_t, 
data_use <- subset(data_use, select=-c(GPPtrend_lm, ERtrend_lm))
#
model <- lm(NEEtrend_lm ~ ., data = data_use)
summary(model)
#
vif_values <- vif(model)
vif_values
 
# adding coexistence rules
sexpr <- expression(!(   (Lgs_t && Lgs_piece_t) 
                         # (TA_gs_t && VPD_gs_t) ||  (TA_ngs_t && VPD_ngs_t) ||
                         # (TA_t && TA_ngs_t) || (TA_t && TA_gs_t) ||
                         # (P_t && P_ngs_t) || (P_t && P_gs_t) ||
                         # (VPD_t && VPD_ngs_t) || (VPD_t && VPD_gs_t) ||
                         # (SW_t && SW_ngs_t) || (SW_t && SW_gs_t)
))
#                         (TA_t && VPD_t) || 
# these pairs cannot coexist 
system.time(dd <- dredge(model, rank = "AICc"))    # , cluster = cl, , subset = sexpr
dd

# P_gs_t and VPD_ngs_t; these two are positive. Grassland is more positive. 
model_best <- lm(NEEtrend_lm ~ P_gs_t + VPD_ngs_t + IGBP.x + Height, data = data_dry)
summary(model_best)

# here VPD trend in growing season is similar to VPD in non-growing season, so this is overall trend; warming-induced VPD increase may enhance CO2 uptake if there is water during growing season.
data_dry %>% 
  ggplot(aes(x=VPD_ngs_t, y=NEEtrend_lm)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green') 

# one LE outliers should be examined!
# this is such a good relationship between VPD_ngs_t and NEE trend. 

# Nmean and Ntrend_lm are not important for NEE trends in dry climates. 


```

```{r trend driver for wet climate }
# identify trend drivers for dry climate
data_wet <- sub_time_series %>% left_join(sites_long[, c("site_ID", "category")], by="site_ID") %>% filter(category=='wet') 
data <- data_wet %>% mutate(Latabs = abs(Lat)) %>%
        dplyr::select(c("NEEtrend_lm", "GPPtrend_lm", "ERtrend_lm", 
                            "TA_t", "P_t", "VPD_t", "SW_t", 'LE_t',
                            "Height", "Nmean"))  # , "Latabs" because not enough meaning. "Nmean"

rcorr_results <- rcorr(as.matrix(data))
cor_matrix <- rcorr_results$r
p_matrix <- rcorr_results$P

# Print correlation matrix
print(cor_matrix)

# Print p-value matrix
print(p_matrix)

#
data_use <- na.omit(data)
#
M=cor(data_use)
testRes = cor.mtest(data_use, conf.level = 0.95)
#
plot <- corrplot(M, p.mat = testRes$p, method = 'circle', insig='blank',
         diag = FALSE)
plot
#
# the impact of Nitrogen deposition on NEE trend
# such a big impact of Nitrogen deposition on NEE trend in wet climate. I cannot believe this. 
#-----------------------------------IMPORTANT-----------------------------------
plot(data_use$NEEtrend_lm, data_use$Nmean)
data_use %>% # filter(Nmean < 30) %>% 
  ggplot(aes(x=Nmean, y=NEEtrend_lm)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')
#-----------------------------------IMPORTANT-----------------------------------

# Nitrogen deposition trend is not that important. likely because the trend data is not very reliable. 
data_use %>% 
  ggplot(aes(x=Nmean, y=NEEtrend_lm)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green')


options(na.action = na.fail)
#
# do not differentiate between growing and non-growing season
data_use <- subset(data_use, select=-c(GPPtrend_lm, ERtrend_lm))
#
model <- lm(NEEtrend_lm ~ ., data = data_use)
summary(model)
#
vif_values <- vif(model)
vif_values
 
# adding coexistence rules
sexpr <- expression(!(   (Lgs_t && Lgs_piece_t) || 
                         (TA_t && TA_ngs_t) || (TA_t && TA_gs_t) ||
                         (P_t && P_ngs_t) || (P_t && P_gs_t) ||
                         (VPD_t && VPD_ngs_t) || (VPD_t && VPD_gs_t) ||
                         (SW_t && SW_ngs_t) || (SW_t && SW_gs_t) ||
                         (LE_t && LE_ngs_t) || (LE_t && LE_gs_t) ||
                         (Nmean && Ntrend_lm)  ))
# these pairs cannot coexist 

system.time(dd <- dredge(model, rank = "AICc"))    # subset = sexpr, cluster = cl
dd
#
model_best <- lm(NEEtrend_lm ~ LE_t + IGBP, data = data_wet)
summary(model_best)
# Evergreen forests and Savahna have more negative values NEE trends. 

model_best <- lm(NEEtrend_lm ~ LE_t + SW_t + IGBP, data = data_wet)
summary(model_best)

```


```{r how about differentiate vegetation types}
data_ENF_cold <- sub_time_series %>% left_join(sites_long[, c("site_ID", "category")], by="site_ID") %>% 
  filter(IGBP %in% c('ENF', "EBF") & category %in% c("cold")) 

data <- data_ENF_cold %>% 
        dplyr::select(c("NEEtrend_lm", "GPPtrend_lm", "ERtrend_lm", 
                            "TA_gs_t", "P_gs_t", "VPD_gs_t", "SW_gs_t", "LE_gs_t", 
                            "TA_ngs_t", "P_ngs_t", "VPD_ngs_t", "SW_ngs_t", "LE_ngs_t", 
                            "Lgs_t", "MAT", "MAP", "Height", "Nmean", "Age"))

rcorr_results <- rcorr(as.matrix(data))
cor_matrix <- rcorr_results$r
p_matrix <- rcorr_results$P

# Print correlation matrix
print(cor_matrix)

# Print p-value matrix
print(p_matrix)

data_use <- na.omit(data)
#
M=cor(data_use)
testRes = cor.mtest(data_use, conf.level = 0.95)
#
plot <- corrplot(M, p.mat = testRes$p, method = 'circle', insig='blank',
         diag = FALSE)
plot

options(na.action = na.fail)
#
# do not differentiate between growing and non-growing season
data_use <- subset(data_use, select=-c(GPPtrend_lm, ERtrend_lm))
#
model <- lm(NEEtrend_lm ~ ., data = data_use)
summary(model)
#
vif_values <- vif(model)
vif_values

system.time(dd <- dredge(model, rank = "AICc"))    # subset = sexpr, cluster = cl
dd

model_best <- lm(NEEtrend_lm ~ Age + MAT + P_gs_t + SW_gs_t + P_ngs_t + VPD_ngs_t, data = data_use)
summary(model_best)

data_use %>% 
  ggplot(aes(x=P_gs_t, y=NEEtrend_lm)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green') 


```

```{r how about differentiate vegetation types}
data_DBF_cold <- sub_time_series %>% left_join(sites_long[, c("site_ID", "category")], by="site_ID") %>% 
  filter(IGBP %in% c('MF', "DBF") ) %>% filter(category %in% c("cold"))
data <- data_DBF_cold %>%
        dplyr::select(c("NEEtrend_lm", "GPPtrend_lm", "ERtrend_lm", 
                            "TA_gs_t", "P_gs_t", "VPD_gs_t", "SW_gs_t", 
                            "TA_ngs_t", "P_ngs_t", "VPD_ngs_t", "SW_ngs_t", 
                            "Lgs_t", "MAT", "MAP", "Height", "Nmean", "Age"))

rcorr_results <- rcorr(as.matrix(data))
cor_matrix <- rcorr_results$r
p_matrix <- rcorr_results$P

# Print correlation matrix
print(cor_matrix)

# Print p-value matrix
print(p_matrix)

data_use <- na.omit(data)
#
M=cor(data_use)
testRes = cor.mtest(data_use, conf.level = 0.95)
#
plot <- corrplot(M, p.mat = testRes$p, method = 'circle', insig='blank',
         diag = FALSE)
plot

options(na.action = na.fail)
#
# do not differentiate between growing and non-growing season
data_use <- subset(data_use, select=-c(GPPtrend_lm, ERtrend_lm))
#
model <- lm(NEEtrend_lm ~ ., data = data_use)
summary(model)
#
vif_values <- vif(model)
vif_values

system.time(dd <- dredge(model, rank = "AICc"))    # subset = sexpr, cluster = cl
dd

model_best <- lm(NEEtrend_lm ~ Age + P_ngs_t + VPD_ngs_t + VPD_gs_t, data = data_use)
summary(model_best)

data_DBF_cold %>% 
  ggplot(aes(x=LE_gs_t, y=NEEtrend_lm)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green') 

```

```{r how about differentiate vegetation types}
data_cold_nonforest <- sub_time_series %>% left_join(sites_long[, c("site_ID", "category")], by="site_ID") %>% 
  filter(!IGBP %in% c('MF', "DBF", "EBF", "ENF") ) %>% filter(category %in% c("cold")) 
data <- data_cold_nonforest %>%
        dplyr::select(c("NEEtrend_lm", "GPPtrend_lm", "ERtrend_lm", 
                            "TA_gs_t", "P_gs_t", "VPD_gs_t", "SW_gs_t", 
                            "TA_ngs_t", "P_ngs_t", "VPD_ngs_t", "SW_ngs_t", 
                            "Lgs_t", "MAT", "MAP", "Nmean"))

rcorr_results <- rcorr(as.matrix(data))
cor_matrix <- rcorr_results$r
p_matrix <- rcorr_results$P

# Print correlation matrix
print(cor_matrix)

# Print p-value matrix
print(p_matrix)

data_use <- na.omit(data)
#
M=cor(data_use)
testRes = cor.mtest(data_use, conf.level = 0.95)
#
plot <- corrplot(M, p.mat = testRes$p, method = 'circle', insig='blank',
         diag = FALSE)
plot

options(na.action = na.fail)
#
# do not differentiate between growing and non-growing season
data_use <- subset(data_use, select=-c(GPPtrend_lm, ERtrend_lm))
#
model <- lm(NEEtrend_lm ~ ., data = data_use)
summary(model)
#
vif_values <- vif(model)
vif_values

system.time(dd <- dredge(model, rank = "AICc"))    # subset = sexpr, cluster = cl
dd

model_best <- lm(NEEtrend_lm ~ MAP + Nmean , data = data_cold_nonforest)
summary(model_best)

data %>% 
  ggplot(aes(x=Nmean, y=NEEtrend_lm)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green') 

data %>% 
  ggplot(aes(x=MAP, y=NEEtrend_lm)) + 
  geom_point() +
  geom_smooth(method = lm) +
  stat_cor(color='green') 

```
