==> scripts/03_01_upscale_global_NEE_trends.Rmd <==
---
title: "identify trend drivers using causal inference"
output: pdf_document
date: "2025-04-08"
author: "Junna Wang"
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```


```{r raw data location}
rm(list=ls())
# change this location where the raw data will be saved accordingly
# dir_rawdata <- '/Users/jw2946/Documents/stability_project/SiteData/'
dir_rawdata <- '/Volumes/MaloneLab/Research/Stability_Project/SiteData'
#
library(librarian)
shelf(tidyverse, ggpubr, terra, ncdf4, future.apply, trend)
#
```


```{r read global vegetation maps}
# linked information about global land_cover_class maps
# https://thredds.daac.ornl.gov/thredds/ncss/grid/ornldaac/968/Land_Cover_Class_0.25degree.nc4/dataset.html
# resolution 0.25 degree
# type 12, croplands; type 13, urban areas; 
# type_1 = Evergreen Needleleaf Forests; type2: Evergreen Broadleaf Forests; type3: Deciduous Needleleaf Forests;  
# type_4 = Deciduous Broadleaf Forests; type_5 = Mixed Forests
# type_11 = Permanent Wetlands; 

```


```{r get start and end of growing season}
data_grow_season <- read.csv('data/growing_season_yearly.csv')
# sub_time_series <- read.csv("data/sub_time_series_potential_drivers_sen.csv")
sites_long <- read.csv('data/long_term_ec_sites.csv')
#
sos_eos <- data_grow_season %>% group_by(site_ID) %>% 
  summarise(sos=mean(sos_TRS+sos_DER, na.rm=T)/2, eos=mean(eos_TRS+eos_DER, na.rm=T)/2) %>% 
  left_join(sites_long[, c('site_ID', 'IGBP', 'Climate.class', 'category')]) %>% 
  mutate(group = case_when(category=='dry' ~ 'dry', category=='wet' ~ 'wet', 
                           category=='cold' & IGBP %in% c('ENF', 'EBF') ~ 'cold_ENF',
                           category=='cold' & IGBP %in% c('DBF', 'MF') ~ 'cold_DBF',
                           category=='cold' & !IGBP %in% c('ENF', 'EBF', 'DBF', 'MF') ~ 'cold_nonforest')) %>%
  arrange(group, Climate.class) %>% 
  filter(site_ID != 'CL-SDF') 

sos_eos_group <- sos_eos %>% group_by(category, Climate.class) %>% 
  summarise(sos_group=round(mean(sos)), eos_group=round(mean(eos))) 
sos_eos_group$sos_group[sos_eos_group$category=='wet'] <- 1
sos_eos_group$eos_group[sos_eos_group$category=='wet'] <- 365
sos_eos_group <- sos_eos_group %>% filter(!(category == 'wet' & Climate.class == "Csa"))

# we add a few climate classes with which we have sites with similar climate
sos_eos_group <- rbind(sos_eos_group, data.frame(category = c("dry", "dry", "cold", "cold", "cold", "cold", "wet", "wet", "wet"),
                                                 Climate.class = c("BWk", "Csc", "Cfc", "Dwa", "Dwb", "Dwd", 'Aw', 'Am', 'Af'), 
                                                 sos_group = c(90, 76, 110, 131, 131, 131, 1, 1, 1), 
                                                 eos_group = c(200, 210, 270, 254, 254, 254, 365, 365, 365)))
sos_eos_group <- sos_eos_group %>% arrange(category, Climate.class)
#
sos_eos_group <- sos_eos_group %>% mutate(climate_id = case_when(
  Climate.class == 'Dwa' ~ 21, Climate.class == 'Dwb' ~ 22,
  Climate.class == 'Dwc' ~ 23, Climate.class == 'Dwd' ~ 24,
  #
  Climate.class == 'Dfa' ~ 25, Climate.class == 'Dfb' ~ 26,
  Climate.class == 'Dfc' ~ 27, Climate.class == 'Dfd' ~ 28,
  Climate.class == 'ET' ~ 29, Climate.class == 'EF' ~ 30,
  #
  Climate.class == 'Csc' ~ 10, Climate.class == 'Csb' ~ 9, Climate.class == 'Csa' ~ 8, 
  Climate.class == 'Bsk' ~ 7, Climate.class == 'Bsh' ~ 6, 
  Climate.class == 'BWk' ~ 5, Climate.class == 'BWh' ~ 4,
  #
  Climate.class == 'Aw' ~ 3, Climate.class == 'Am' ~ 2, Climate.class == 'Af' ~ 1, 
  #
  Climate.class == 'Dsa' ~ 17, Climate.class == 'Dsb' ~ 18, 
  Climate.class == 'Dsc' ~ 19, Climate.class == 'Dsd' ~ 20, 
  #
  Climate.class == 'Cfa' ~ 14, Climate.class == 'Cfb' ~ 15, Climate.class == 'Cfc' ~ 16,
  #
  Climate.class == 'Cwa' ~ 11, Climate.class == 'Cwb' ~ 12, Climate.class == 'Cwc' ~ 13))
#  mutate(zone_id = paste(group, climate_id, sep="-"))

# the climate class we do not have data are 
# Af (rainforest) and Aw (Savvana), all the Dw climates (Continential dry winter climate)
```


```{r get global zonation based on climate and vegetation}
koppen_climate <- terra::rast('data/koppen_geiger_0p00833333.tif')
land_cover <- terra::rast('data/968_Land_Cover_Class_0.25degree.nc')
#
# plot(koppen_climate)
plot(land_cover[[2]]%in% c(4:5))
# large croplands on earth
# unify resolution: 0.25 degree
koppen_climate_0.25 <- terra::resample(koppen_climate, land_cover, method="near")
# plot(koppen_climate_0.25)
#
# a data frame to store zone info
df_zone <- data.frame(koppen_climate=as.integer(terra::values(koppen_climate_0.25)),
                      land_cover=as.integer(terra::values(land_cover[[2]])))
df_zone <- df_zone %>% mutate(group = case_when(
  koppen_climate %in% c(2, 14, 11:13) & !land_cover %in% c(12,13,15,16,0) ~ "wet",  
  koppen_climate %in% c(4:10) & !land_cover %in% c(12,13,15,16,0) ~ "dry",  # remove 0water, 12cropland, 13urban, 15ice, 16barren
  koppen_climate %in% c(15:30) & land_cover %in% c(6:9, 14) ~ "cold_nonforest", 
  koppen_climate %in% c(15:30) & land_cover %in% c(10) ~ "cold_nonforest", 
  koppen_climate %in% c(15:30) & land_cover %in% c(11) ~ "cold_nonforest", 
  koppen_climate %in% c(15:30) & land_cover %in% c(1:3) ~ "cold_ENF",
  koppen_climate %in% c(15:30) & land_cover %in% c(4:5) ~ "cold_DBF"))
# %>% mutate(grow_season_id = ifelse(is.na(koppen_climate) | is.na(group), NA, paste(group, koppen_climate, sep='-')))

zone_grow_season <- koppen_climate_0.25
terra::values(zone_grow_season)[!df_zone$koppen_climate %in% sos_eos_group$climate_id] <- NA   
plot(zone_grow_season)
# why there are so many missing points? these are water areas. 
# we have 20 climate zones, missing Aw, Am, Af, and Dsa, Dsb, Dsc, Dsd

# zone for climate-vegetation groups
zone_group <- koppen_climate_0.25
terra::values(zone_group) <- NA
terra::values(zone_group) <- df_zone$group
plot(zone_group)
writeRaster(zone_group, filename = 'data/zones_climate_vegetation_group.tif', overwrite=TRUE)

#------------------------------------
# calculate zone_grow_season masks
zone_masks <- rast()
for (i in 1:nrow(sos_eos_group)) {
  m <- zone_grow_season == sos_eos_group$climate_id[i]
  terra::values(m)[!terra::values(m)] <- NA
  names(m) <- paste0("zone_", sos_eos_group$climate_id[i])
  zone_masks <- c(zone_masks, m)
}

north_hemispher <- init(zone_masks[[1]], "y") > 0
south_hemispher <- init(zone_masks[[1]], "y") <= 0

zone_masks_north <- mask(zone_masks, north_hemispher, maskvalues=0)
zone_masks_south <- mask(zone_masks, south_hemispher, maskvalues=0)

```


```{r a function to calculate growing season mean by zones}

process_one_year <- function(r, zone_masks_north, zone_masks_south, sos_eos_group, gs=TRUE) {
  output <- terra::rast(r[[1]])
  terra::values(output) <- NA
  
  for (rid in 1:nrow(sos_eos_group)) {
#    browser()
    days_north <- sos_eos_group$sos_group[rid]:sos_eos_group$eos_group[rid]
    days_south <- ifelse(days_north + 182 > 365, days_north + 182 - 365, days_north + 182)
    
    if (!gs) {
      days_north <- setdiff(1:365, days_north)
      days_south <- setdiff(1:365, days_south)
    }
    
    # Subset days for the region
    if (length(days_north) == 0) {next}
    r_north <- r[[days_north]]
    r_south <- r[[days_south]]

    # Compute mean across time
    # mean_north <- terra::app(r_north, mean, na.rm = TRUE, cores=8)
    # mean_south <- terra::app(r_south, mean, na.rm = TRUE, cores=8)
    # 
    # # mask results; this is faster than mask first
    # mean_north <- mask(mean_north, zone_masks_north[[rid]])
    # mean_south <- mask(mean_south, zone_masks_south[[rid]])
    
    # mean_north <- zonal(r_north, zone_masks_north[[rid]], fun = "mean", na.rm = TRUE)
    # mean_south <- zonal(r_south, zone_masks_south[[rid]], fun = "mean", na.rm = TRUE)
    cells_north <- which(!is.na(terra::values(zone_masks_north[[rid]])))
    cells_south <- which(!is.na(terra::values(zone_masks_south[[rid]])))
    #
    if (length(cells_north) > 0) {
      output[cells_north] <- rowMeans(r_north[cells_north], na.rm=T)
    }    
    #
    if (length(cells_south) > 0) {
      output[cells_south] <- rowMeans(r_south[cells_south], na.rm=T)
    }
    #
    # Merge them into final output
    # output <- cover(output, mean_north)
    # output <- cover(output, mean_south)
    # print(paste0('finish zone ', rid))
  }
  return(output)
}

```


```{r read data for calculating growing season mean each year}
# in the southern hemisphere, these date should be the date + 180. 
met_files <- list.files(file.path(dir_rawdata, '/ScaleUp_Global/met_day_2005_2023/'), pattern='.nc$', full.names = T)
#
sw_data <- list()
tam_data <- list()
tp_data <- list()
#
df_file_info <- data.frame(year=integer(), var=character(), file=character())
#
for (i in 1:length(met_files)) {
  nc <- nc_open(met_files[i])
  time_str <- ncatt_get(nc, "valid_time", "units")$value
  year <- as.numeric(sub(".*since (\\d{4})-.*", "\\1", time_str))
  varname <- names(nc$var)[1]
  df_file_info[i,1] <- year
  df_file_info[i,2] <- varname
  df_file_info[i,3] <- met_files[i]
  nc_close(nc)
}
df_file_info <- df_file_info %>% arrange(var, year)

# it takes 3-4 hours to rotate these rasters.
# work on t2m
t2m_data <- list()
df_file_info_subset <- subset(df_file_info, var=='t2m')    # nrow(df_file_info_subset)
for (i in 1:nrow(df_file_info_subset)) {
  print(i)
  met_data <- terra::rast(df_file_info_subset$file[i])
  met_data <- terra::rotate(met_data)
  t2m_data <- c(t2m_data, met_data)
}

tp_data <- list()
df_file_info_subset <- subset(df_file_info, var=='tp')
for (i in 1:nrow(df_file_info_subset)) {
  print(i)
  met_data <- terra::rast(df_file_info_subset$file[i])
  met_data <- terra::rotate(met_data)
  tp_data <- c(tp_data, met_data)
}

ssrd_data <- list()
df_file_info_subset <- subset(df_file_info, var=='ssrd')
for (i in 1:nrow(df_file_info_subset)) {
  print(i)
  met_data <- terra::rast(df_file_info_subset$file[i])
  met_data <- terra::rotate(met_data)
  ssrd_data <- c(ssrd_data, met_data)
}

```


```{r call function to calculate growing season mean}
# call function: this will take 3-4 hours
# test <- process_one_year(t2m_data[[1]], zone_masks_north, zone_masks_south, sos_eos_group)
#
# resample masks to be consistent with nc data
zone_masks_north <- terra::resample(zone_masks_north, t2m_data[[1]], method = "near")
zone_masks_south <- terra::resample(zone_masks_south, t2m_data[[1]], method = "near")


t2m_gs <- rast()  # length(t2m_data)
for (i in 1:length(t2m_data)) {
  test <- process_one_year(t2m_data[[i]], zone_masks_north, zone_masks_south, sos_eos_group, gs=TRUE)
  t2m_gs <- c(t2m_gs, test)
  print(i)
}
#
writeRaster(t2m_gs, filename = 'data/t2m_gs_2005_2023.tif', overwrite=TRUE)

tp_gs <- rast()
for (i in 1:length(tp_data)) {
  test <- process_one_year(tp_data[[i]], zone_masks_north, zone_masks_south, sos_eos_group, gs=TRUE)
  tp_gs <- c(tp_gs, test)
  print(i)
}
#
writeRaster(tp_gs, filename = 'data/tp_gs_2005_2023.tif', overwrite=TRUE)

# non-growing season
tp_ngs <- rast()
for (i in 1:length(tp_data)) {
  test <- process_one_year(tp_data[[i]], zone_masks_north, zone_masks_south, sos_eos_group, gs=FALSE)
  tp_ngs <- c(tp_ngs, test)
  print(i)
}
#
writeRaster(tp_ngs, filename = 'data/tp_ngs_2005_2023.tif', overwrite=TRUE)


ssrd_gs <- rast()
for (i in 1:length(tp_data)) {
  test <- process_one_year(ssrd_data[[i]], zone_masks_north, zone_masks_south, sos_eos_group, gs=TRUE)
  ssrd_gs <- c(ssrd_gs, test)
  print(i)
}
#
writeRaster(ssrd_gs, filename = 'data/ssrd_gs_2005_2023.tif', overwrite=TRUE)

```


```{r deal with ET data from FLUXCOM in the past}
ET_files <- list.files(file.path(dir_rawdata, '/ScaleUp_Global/et_day_2005_2021/'), pattern='.nc$', full.names = T)
et_data <- list()
for (i in 1:length(ET_files)) {
  print(i)
  met_data <- terra::rast(ET_files[i])
  # plot(met_data)
  et_data <- c(et_data, met_data)
}

# resample masks to be consistent with nc data
zone_masks_north <- resample(zone_masks_north, met_data, method = "near")
zone_masks_south <- resample(zone_masks_south, met_data, method = "near")

# calculate growing season average;
et_gs <- rast()  # length(t2m_data)
for (i in 1:length(et_data)) {
  test <- process_one_year(et_data[[i]], zone_masks_north, zone_masks_south, sos_eos_group)
  et_gs <- c(et_gs, test)
  print(i)
}
writeRaster(et_gs, filename = 'data/et_gs_2005_2023.tif', overwrite=TRUE)


```


```{r calculate past NEE trends}
t2m_gs <- terra::rast('data/t2m_gs_2005_2023.tif')
tp_gs <- terra::rast('data/tp_gs_2005_2023.tif')
tp_ngs <- terra::rast('data/tp_ngs_2005_2023.tif')
ssrd_gs <- terra::rast('data/ssrd_gs_2005_2023.tif')
et_gs <- terra::rast('data/et_gs_2005_2023.tif')
#
plot(t2m_gs)
# multi-cores do not work for this case
t2m_gs_slope <- terra::app(t2m_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
mean(terra::values(t2m_gs_slope), na.rm=T)   # ~ 0.04757178

# t2m_gs_pvalue <- terra::app(t2m_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$p.value)})
# sum(terra::values(t2m_gs_pvalue) < 0.05, na.rm=T) / sum(!is.na(terra::values(t2m_gs_pvalue)))
# 0.1960132 significant trends at 0.05; 0.3176152 significant trends at 0.1;

tp_gs_slope <- terra::app(tp_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
mean(terra::values(tp_gs_slope), na.rm=T)   # -0.007202292

# tp_gs_pvalue <- terra::app(tp_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$p.value)})
# sum(terra::values(tp_gs_pvalue) < 0.1, na.rm=T) / sum(!is.na(terra::values(tp_gs_pvalue)))
# 0.06254149 significant trends at 0.05; 0.1174835 significant trends at 0.1; similar to our dataset!

tp_ngs_slope <- terra::app(tp_ngs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
mean(terra::values(tp_ngs_slope), na.rm=T)   # 0.01022895; precipitation in non-growing season is increasing more. 

# tp_ngs_pvalue <- terra::app(tp_ngs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$p.value)})
# sum(terra::values(tp_ngs_pvalue) < 0.05, na.rm=T) / sum(!is.na(terra::values(tp_ngs_pvalue)))   
# 0.07816733 significant trends at 0.05; 0.1347128 significant trends at 0.1; similar to our dataset!

##
ssrd_gs_slope <- terra::app(ssrd_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
mean(terra::values(ssrd_gs_slope), na.rm=T)   # 0.006198188; this is so small!

ssrd_gs_pvalue <- terra::app(ssrd_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$p.value)})
sum(terra::values(ssrd_gs_pvalue) < 0.05, na.rm=T) / sum(!is.na(terra::values(ssrd_gs_pvalue)))
# 0.09116757 significant trends at 0.05; 0.1680927 significant trends at 0.1;

et_gs_slope <- terra::app(et_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
mean(terra::values(et_gs_slope), na.rm=T)   # 0.02263143; this is so small!
writeRaster(et_gs_slope, 'data/et_gs_slope_FLUXCOM_2005_2023.tif', overwrite=T)

# et_gs_pvalue <- terra::app(et_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$p.value)})
# sum(terra::values(et_gs_pvalue) < 0.1, na.rm=T) / sum(!is.na(terra::values(et_gs_pvalue))) 
# 0.1348455 significant trends at 0.05; 0.1680927 significant trends at 0.2146213;

# read canopy height
canopy_height <- terra::rast('data/global_canopy_height_2020.tif')

# align up all the slope result and trend result
t2m_gs_slope <- terra::resample(t2m_gs_slope, et_gs_slope)
tp_gs_slope <- terra::resample(tp_gs_slope, et_gs_slope)
tp_ngs_slope <- terra::resample(tp_ngs_slope, et_gs_slope)
ssrd_gs_slope <- terra::resample(ssrd_gs_slope, et_gs_slope)
zone_group <- terra::resample(zone_group, et_gs_slope, method='near')
canopy_height <- terra::resample(canopy_height, et_gs_slope)

# predicting NEE trend for each climate-vegetation class
NEE_trend_past <- zone_group
terra::values(NEE_trend_past) <- NA
unique_climate_veg_groups <- levels(zone_group)[[1]]
for (group in unique_climate_veg_groups$label) {
  cells <- which(terra::values(zone_group) == unique_climate_veg_groups$value[unique_climate_veg_groups$label==group])
  if (group == 'wet') {
    NEE_trend_past[cells] <- 0.60216 * ssrd_gs_slope[cells] + -0.00942
  } else if (group == 'dry') {
    NEE_trend_past[cells] <- + 0.05882 + 0.92125 * tp_ngs_slope[cells] - 1.21953 * tp_gs_slope[cells]  # non-growing season precipitation
  } else if (group == 'cold_ENF') {
    NEE_trend_past[cells] <- -0.74007 * tp_gs_slope[cells] + -0.67349 * (-0.03637 - 0.30871 * tp_gs_slope[cells] - 0.25552 * t2m_gs_slope[cells]) + -0.06156
  } else if (group == 'cold_DBF') {
    NEE_trend_past[cells] <- -0.36844 * et_gs_slope[cells] + 0.01111   # 
  } else if (group %in% c('cold_nonforest')) {
    NEE_trend_past[cells] <- -0.292777 * tp_gs_slope[cells] + 0.457980 * ssrd_gs_slope[cells] + -0.011463 * canopy_height[cells] + 0.135375
  }
  #
  print(mean(terra::values(NEE_trend_past)[cells], na.rm=T))
  numbers <- terra::values(NEE_trend_past)[cells]
  df <- data.frame(value=numbers[!is.na(numbers)])
  plot <- ggplot(df, aes(x = value)) +
  geom_density(fill = "skyblue", alpha = 0.5) +
  labs(title = group, x = "NEE trend", y = "Density")
  print(plot)
  
  # Are driver trends cancelled out at large scales?
  print(paste0("Past mean prep trend in a group: ", group, mean(terra::values(tp_gs_slope)[cells], na.rm=T)))
  print(paste0("Past mean ssrd trend in a group: ", group, mean(terra::values(ssrd_gs_slope)[cells], na.rm=T)))
  print(paste0("Past mean t2m trend in a group: ", group, mean(terra::values(t2m_gs_slope)[cells], na.rm=T)))
}
# [1] 0.0006620509
# [1] -0.02179434
# [1] 0.07309876
# [1] 0.07049765
# [1] -0.01247933

plot(NEE_trend_past)
mean(terra::values(NEE_trend_past), na.rm=T)  # 0.03933005

writeRaster(NEE_trend_past, filename = 'data/NEE_trend_past_prediction.tif', overwrite=TRUE)
# [1] "Past mean prep trend in a group: cold_DBF-0.0106894772280376"
# [1] "Past mean ssrd trend in a group: cold_DBF0.0174385703532912"
# [1] "Past mean t2m trend in a group: cold_DBF0.0372846465169043"
# [1] -0.02179434
# [1] "Past mean prep trend in a group: cold_ENF-0.0130431158948662"
# [1] "Past mean ssrd trend in a group: cold_ENF0.0282002007813707"
# [1] "Past mean t2m trend in a group: cold_ENF0.0484039722938379"
# [1] 0.07309876
# [1] "Past mean prep trend in a group: cold_nonforest-0.000739786467414556"
# [1] "Past mean ssrd trend in a group: cold_nonforest0.0142696955056926"
# [1] "Past mean t2m trend in a group: cold_nonforest0.0537303760982132"
# [1] 0.07049765
# [1] "Past mean prep trend in a group: dry-0.0134900779262378"
# [1] "Past mean ssrd trend in a group: dry0.00419883350499074"
# [1] "Past mean t2m trend in a group: dry0.0364109797163263"
# [1] -0.01247933
# [1] "Past mean prep trend in a group: wet-0.00376746135272869"
# [1] "Past mean ssrd trend in a group: wet-0.00508059767128605"
# [1] "Past mean t2m trend in a group: wet0.0707241803036828"
# [1] 0.03933005


```

```{r get future climate average during growing and non-growing season}
# get the average of the three global circulation models: GFDL, MICRO6, and MPI-ESM
vars <- c('pr', 'rsds', 'tas')
for (var in vars) {
  folder <- paste0(var, '_day_2026_2050')
  #
  nc_files <- list.files(file.path(dir_rawdata, 'ScaleUp_Global', folder), pattern='.nc$', full.names = T)
  tp_data_fut3 <- list()
  for (i in 1:length(nc_files)) {
    met_data <- terra::rast(nc_files[i])
    tp_data_fut3 <- c(tp_data_fut3, met_data)
  }
  
  tp_data_fut <- list()
  nyears <- length(nc_files) / 3
  # we only used the data between 2026 and 2045
  for (i in 1:20) {
  # for (i in 1:nyears) {
    tp_data_fut <- c(tp_data_fut, (tp_data_fut3[[i]] + tp_data_fut3[[i+nyears]] + tp_data_fut3[[i+nyears*2]]) / 3)
    print(i)
  }
  
  # resample masks to be consistent with nc data
  test <- terra::rotate(tp_data_fut[[1]])
  zone_masks_north <- resample(zone_masks_north, test, method = "near")
  zone_masks_south <- resample(zone_masks_south, test, method = "near")
  
  # average tp of growing season and non-growing season
  tp_gs_fut <- rast()
  for (i in 1:length(tp_data_fut)) {
    tp_data_fut[[i]] <- terra::rotate(tp_data_fut[[i]])
    test <- process_one_year(tp_data_fut[[i]], zone_masks_north, zone_masks_south, sos_eos_group, gs=TRUE)
    tp_gs_fut <- c(tp_gs_fut, test)
    print(i)
  }
  writeRaster(tp_gs_fut, filename = paste0('data/', var, '_gs_2026_2050.tif'), overwrite=TRUE)
  
  # non-growing season
  if (var == 'pr') {
    tp_ngs_fut <- rast()
    for (i in 1:length(tp_data_fut)) {
      test <- process_one_year(tp_data_fut[[i]], zone_masks_north, zone_masks_south, sos_eos_group, gs=FALSE)
      tp_ngs_fut <- c(tp_ngs_fut, test)
      print(i)
    }
    writeRaster(tp_ngs_fut, filename = paste0('data/', var, '_ngs_2026_2050.tif'), overwrite=TRUE)      
  }
}

```


```{r get change of ET from coarse global climate models}
# I want to get the change of ET from 2015-2025 to 2026-2050. 
# assuming growing season is 5-9; 
# calculate changes in slopes using ET projections from 6 global climate models under ssp245;
# BCC-CSM2-MR, CAMS-CSM1-0, CAS-ESM2-0, FIO-ESM-2-0, MIROC6, CESM2_ssp245
folder <- 'et_day_2026_2050'
nc_files <- list.files(file.path(dir_rawdata, 'ScaleUp_Global', folder), pattern='.nc$', full.names = T)
et_data_fut3 <- list()
for (i in 1:length(nc_files)) {
  met_data <- terra::rast(nc_files[i])
  et_data_fut3 <- c(et_data_fut3, met_data)
}
et_data_fut3[[6]] <- c(et_data_fut3[[6]], et_data_fut3[[7]])

nmodels <- 6

et_slope_model_past <- list()
et_slope_model_fut <- list()
et_slope_model_past_fut_change <- list()
for (imodel in 1:nmodels) {
  et_data_year <- rast()
  data_north <- terra::mask(et_data_fut3[[imodel]], init(et_data_fut3[[imodel]], "y") > 0, maskvalues=0)
  data_south <- terra::mask(et_data_fut3[[imodel]], init(et_data_fut3[[imodel]], "y") <= 0, maskvalues=0)
  #
  for (iyear in 1:36) {
    r_north <- terra::app(data_north[[c(5:9) + (iyear-1)*12]], fun=mean, na.rm=T)
    r_south <- terra::app(data_south[[c(1:3, 11:12) + (iyear-1)*12]], fun=mean, na.rm=T)
    et_data_year <- c(et_data_year, terra::cover(r_north, r_south))
    print(iyear)
  }
  # 2015-2030
  et_slope_model_past[[imodel]] <- terra::app(et_data_year[[1:16]], function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
#  plot(et_slope_model_past[[imodel]])
  et_slope_model_past[[imodel]] <- terra::rotate(et_slope_model_past[[imodel]])
  # 2031-2046
  et_slope_model_fut[[imodel]] <- terra::app(et_data_year[[17:32]], function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
  et_slope_model_fut[[imodel]] <- terra::rotate(et_slope_model_fut[[imodel]])
}

# Compare the past ET trends from FLUXCOM with the current ET trends from climate models
et_gs_slope_FLUXCOM <- terra::rast('data/et_gs_slope_FLUXCOM_2005_2023.tif')
plot(et_gs_slope_FLUXCOM)
for (imodel in 1:nmodels) {
  et_slope_model_past[[imodel]] <- terra::resample(et_slope_model_past[[imodel]], et_gs_slope_FLUXCOM)
  et_slope_model_past[[imodel]] <- terra::mask(et_slope_model_past[[imodel]], et_gs_slope_FLUXCOM)
  plot(et_slope_model_past[[imodel]])
  print(cor.test(terra::values(et_gs_slope_FLUXCOM), terra::values(et_slope_model_past[[imodel]])))
  #
  et_slope_model_fut[[imodel]] <- terra::resample(et_slope_model_fut[[imodel]], et_gs_slope_FLUXCOM)
  et_slope_model_fut[[imodel]] <- terra::mask(et_slope_model_fut[[imodel]], et_gs_slope_FLUXCOM)
}

et_slope_model_past <- terra::app(terra::rast(et_slope_model_past), fun='mean', na.rm=T)
et_slope_model_fut  <- terra::app(terra::rast(et_slope_model_fut), fun='mean', na.rm=T)

print(cor.test(terra::values(et_gs_slope_FLUXCOM), terra::values(et_slope_model_past)))
# r = 0.20099 and significant


et_slope_model_past_fut_change <- et_slope_model_fut - et_slope_model_past
plot(et_slope_model_past_fut_change)
#
writeRaster(et_slope_model_past_fut_change, filename = 'data/et_slope_past_fut_change.tif', overwrite=TRUE)
# 
```

```{r predict for future NEE trend}
t2m_gs <- terra::rast('data/tas_gs_2026_2050.tif')
tp_gs <- terra::rast('data/pr_gs_2026_2050.tif')
tp_ngs <- terra::rast('data/pr_ngs_2026_2050.tif')
ssrd_gs <- terra::rast('data/rsds_gs_2026_2050.tif')   
et_gs <- terra::rast('data/et_gs_2005_2023.tif')       # this is in the past
#
et_slope_model_past_fut_change <- terra::rast('data/et_slope_past_fut_change.tif')

#-----calculate slopes
t2m_gs_slope <- terra::app(t2m_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
mean(terra::values(t2m_gs_slope), na.rm=T)   # ~ 0.04976986; similar to past climate data

# t2m_gs_pvalue <- terra::app(t2m_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$p.value)})
# sum(terra::values(t2m_gs_pvalue) < 0.05, na.rm=T) / sum(!is.na(terra::values(t2m_gs_pvalue)))
# 0.2675935 significant trends at 0.05; 

tp_gs_slope <- terra::app(tp_gs, function(x) {ifelse(any(is.na(x)), NA, ifelse(sd(x)==0, 0, sens.slope(x)$estimate / sd(x)))})
mean(terra::values(tp_gs_slope), na.rm=T)   # 0.002988621

# tp_gs_pvalue <- terra::app(tp_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$p.value)})
# sum(terra::values(tp_gs_pvalue) < 0.05, na.rm=T) / sum(!is.na(terra::values(tp_gs_pvalue)))
# 0.05651056 significant trends at 0.05; 

tp_ngs_slope <- terra::app(tp_ngs, function(x) {ifelse(any(is.na(x)), NA, ifelse(sd(x)==0, 0, sens.slope(x)$estimate / sd(x)))})
mean(terra::values(tp_ngs_slope), na.rm=T)   # -0.001253997

# tp_ngs_pvalue <- terra::app(tp_ngs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$p.value)})
# sum(terra::values(tp_ngs_pvalue) < 0.05, na.rm=T) / sum(!is.na(terra::values(tp_ngs_pvalue)))
# 0.05474039 significant trends at 0.05; 

##
ssrd_gs_slope <- terra::app(ssrd_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
mean(terra::values(ssrd_gs_slope), na.rm=T)   # 0.002495198;

# ssrd_gs_pvalue <- terra::app(ssrd_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$p.value)})
# sum(terra::values(ssrd_gs_pvalue) < 0.05, na.rm=T) / sum(!is.na(terra::values(ssrd_gs_pvalue)))
# 0.06339896 significant trends at 0.05; 

et_gs_FLUXCOM_slope <- terra::app(et_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
mean(terra::values(et_gs_FLUXCOM_slope), na.rm=T)   # 0.02280891; this is so small!

# add changes in et slope
et_gs_slope <- et_gs_FLUXCOM_slope + et_slope_model_past_fut_change
plot(et_gs_slope)

# read canopy height
canopy_height <- terra::rast('data/global_canopy_height_2020.tif')

# align up all the slope result and trend result
t2m_gs_slope <- terra::resample(t2m_gs_slope, et_gs_slope)
tp_gs_slope <- terra::resample(tp_gs_slope, et_gs_slope)
tp_ngs_slope <- terra::resample(tp_ngs_slope, et_gs_slope)
ssrd_gs_slope <- terra::resample(ssrd_gs_slope, et_gs_slope)
zone_group <- terra::resample(zone_group, et_gs_slope, method='near')
canopy_height <- terra::resample(canopy_height, et_gs_slope)

# predicting NEE trend for each climate-vegetation class
NEE_trend_fut <- zone_group
terra::values(NEE_trend_fut) <- NA
unique_climate_veg_groups <- levels(zone_group)[[1]]
for (group in unique_climate_veg_groups$label) {
  cells <- which(terra::values(zone_group) == unique_climate_veg_groups$value[unique_climate_veg_groups$label==group])
  if (group == 'wet') {
    NEE_trend_fut[cells] <- 0.60216 * ssrd_gs_slope[cells] + -0.00942
  } else if (group == 'dry') {
    NEE_trend_fut[cells] <- + 0.05882 + 0.92125 * tp_ngs_slope[cells] - 1.21953 * tp_gs_slope[cells]  # non-growing season precipitation
  } else if (group == 'cold_ENF') {
    NEE_trend_fut[cells] <- -0.74007 * tp_gs_slope[cells] + -0.67349 * (-0.03637 - 0.30871 * tp_gs_slope[cells] - 0.25552 * t2m_gs_slope[cells]) + -0.06156
    # our own research
    # assuming no growing season length trend
    # NEE_trend_fut[cells] <- -0.74007 * tp_gs_slope[cells] + -0.67349 * (-0.03590 - 0.23430 * t2m_gs_slope[cells]) + -0.06156
  } else if (group == 'cold_DBF') {
    NEE_trend_fut[cells] <- -0.36844 * et_gs_slope[cells] + 0.01111   # 
  } else if (group == 'cold_nonforest') {
    NEE_trend_fut[cells] <- -0.292777 * tp_gs_slope[cells] + 0.457980 * ssrd_gs_slope[cells] + -0.011463 * canopy_height[cells] + 0.135375
  }
  #
  print(mean(terra::values(NEE_trend_fut)[cells], na.rm=T))
  numbers <- terra::values(NEE_trend_fut)[cells]
  df <- data.frame(value=numbers[!is.na(numbers)])
  plot <- ggplot(df, aes(x = value)) +
  geom_density(fill = "skyblue", alpha = 0.5) +
  labs(title = group, x = "NEE trend", y = "Density")
  print(plot)
  # Are driver trends cancelled out at large scales?
  print(paste0("Future mean prep trend in a group: ", group, mean(terra::values(tp_gs_slope)[cells], na.rm=T)))
  print(paste0("Future mean ssrd trend in a group: ", group, mean(terra::values(ssrd_gs_slope)[cells], na.rm=T)))
  print(paste0("Future mean t2m trend in a group: ", group, mean(terra::values(t2m_gs_slope)[cells], na.rm=T)))
}
plot(NEE_trend_fut)
mean(terra::values(NEE_trend_fut), na.rm=T)  # 0.03079345; I should use area averaged values. 

writeRaster(NEE_trend_fut, filename = 'data/NEE_trend_future_prediction.tif', overwrite=TRUE)

# [1] 0.001524375
# [1] -0.04325326
# [1] 0.05841176       # this one is smaller, this is non-forests. 
# [1] 0.06166086
# [1] -0.0006226519     # this one changed its sign. 

# [1] -0.04325326
# [1] "Future mean prep trend in a group:cold_ENF0.0199694059078339"
# [1] "Future mean ssrd trend in a group:cold_ENF-0.0166143229967202"
# [1] "Future mean t2m trend in a group:cold_ENF0.0257932349937162"
# [1] 0.05867833
# [1] "Future mean prep trend in a group:cold_nonforest0.0151858063888047"
# [1] "Future mean ssrd trend in a group:cold_nonforest-0.00282362432768481"
# [1] "Future mean t2m trend in a group:cold_nonforest0.0496374942133618"
# [1] 0.06166086
# [1] "Future mean prep trend in a group:dry-0.00801849479131112"
# [1] "Future mean ssrd trend in a group:dry0.00467357055481933"
# [1] "Future mean t2m trend in a group:dry0.0504363448075233"
# [1] -0.0006226519
# [1] "Future mean prep trend in a group:wet0.00656352315517214"
# [1] "Future mean ssrd trend in a group:wet0.0146096521058984"
# [1] "Future mean t2m trend in a group:wet0.0616470550838771"



```


```{r standard deviation of NEE}
#
years <- 2005:2021
folder1 <- 'nee_fluxcom_2005_2021'
folder2 <- paste0('FLUXCOM-X-BASE net ecosystem exchange for ', years)
folder3 <- paste0('NEE_', years, '_025_daily.nc')
#
NEE_annual_2005_2021 <- terra::rast()
for (i in 1:length(years)) {
  print(i)
  nc_file <- list.files(file.path(dir_rawdata, 'ScaleUp_Global', folder1, folder2[i]), pattern='_025_daily.nc$', full.names = T)
  NEE_daily <- terra::rast(nc_file)
  NEE_annual <- terra::app(NEE_daily, fun = 'sum', na.rm=T)
  NEE_annual_2005_2021 <- c(NEE_annual_2005_2021, NEE_annual)
}

writeRaster(NEE_annual_2005_2021, filename = 'data/NEE_annual_2005_2021_fluxcom.tif', overwrite=TRUE)

NEE_sd_2005_2021_fluxcom <- terra::app(NEE_annual_2005_2021, fun = 'sd', na.rm=T)
NEE_slope_2005_2021_fluxcom <- terra::app(NEE_annual_2005_2021, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate)})
plot(NEE_sd_2005_2021_fluxcom)
plot(NEE_slope_2005_2021_fluxcom)

```


```{r convert standardized NEE trends to normal NEE trend}
NEE_annual_2005_2021 <- terra::rast('data/NEE_annual_2005_2021_fluxcom.tif')
NEE_sd_2005_2021_fluxcom <- terra::app(NEE_annual_2005_2021, fun = 'sd', na.rm=T)

# we cannot use this, because so many areas on earth the sd is just zero; this just does not make sense. 
quantile(terra::values(NEE_sd_2005_2021_fluxcom), c(0.99, 0.9, 0.75, 0.5, 0.25, 0.1, 0.01), na.rm=T)

#
NEE_trend_fut <- terra::rast('data/NEE_trend_future_prediction.tif')
NEE_trend_past <- terra::rast('data/NEE_trend_past_prediction.tif')

NEE_trend_fut_unit <- NEE_trend_fut * NEE_sd_2005_2021_fluxcom
NEE_trend_past_unit <- NEE_trend_past * NEE_sd_2005_2021_fluxcom

plot(NEE_trend_fut_unit, range=c(-20, 30))
plot(NEE_trend_past_unit, range=c(-20, 30))

mean(NEE_trend_fut_unit, na.rm=T)
mean(NEE_trend_past_unit, na.rm=T)
density(NEE_trend_fut_unit)
density(NEE_trend_past_unit)

writeRaster(NEE_trend_fut_unit, filename = 'data/NEE_trend_future_prediction_unit.tif', overwrite=TRUE)
writeRaster(NEE_trend_past_unit, filename = 'data/NEE_trend_past_prediction_unit.tif', overwrite=TRUE)

```


```{r change in growing season solar radiaiton and precipitation}
# here I want to address two questions: 
# (1) do model products underestimate growing season shortwave radiations?
# (2) how global shortwave radiation trends are changing?
# (3) how global precipitation trends are changing? 
# (4) how global warming trends are changing?
# answer Q1: 
ssrd_gs <- terra::rast('data/ssrd_gs_2005_2023.tif')
ssrd_gs_slope_past <- terra::app(ssrd_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})

#
sub_time_series <- read.csv("data/sub_time_series_potential_drivers_sen.csv")
sites_long <- read.csv('data/long_term_ec_sites.csv')
if (! "category" %in% names(sub_time_series)) {
  sub_time_series <- sub_time_series %>% 
    left_join(sites_long[, c("site_ID", "category", "Long")], by="site_ID") %>%
    rename("NEE_trend"="NEEtrend_lm")
}

# extract ssrd_gs_slope at observation locations
xy <- sub_time_series[, c("Long", "Lat")]
sub_time_series$ssrd_gs_slope_sites <- terra::extract(ssrd_gs_slope, xy)$lyr.1
cor.test(sub_time_series$ssrd_gs_slope_sites, sub_time_series$SW_gs_t)
# p-value = 0.002496! climate re-analysis data did consist with site-level shortwave radiation change. 

for (group in unique(sub_time_series$category) ) {
  plot(sub_time_series$ssrd_gs_slope_sites[sub_time_series$category==group], 
       sub_time_series$SW_gs_t[sub_time_series$category==group], main = group)
}

# answer Q2:
ssrd_gs <- terra::rast('data/rsds_gs_2026_2050.tif')   
ssrd_gs_slope_fut <- terra::app(ssrd_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
#
ssrd_gs_slope_fut <- terra::resample(ssrd_gs_slope_fut, ssrd_gs_slope_past)
#
ssrd_gs_slope_change <- ssrd_gs_slope_fut - ssrd_gs_slope_past
plot(ssrd_gs_slope_change)
#
writeRaster(ssrd_gs_slope_change, filename = 'data/ssrd_gs_slope_change.tif', overwrite=TRUE)
# this should be a map of global coverage. 
mean(terra::values(ssrd_gs_slope_change), na.rm=T) # -0.00193066

# answer Q3:
tp_gs <- terra::rast('data/tp_gs_2005_2023.tif')   
tp_gs_slope_past <- terra::app(tp_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
#
tp_gs <- terra::rast('data/pr_gs_2026_2050.tif')   
tp_gs_slope_fut <- terra::app(tp_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
#
tp_gs_slope_fut <- terra::resample(tp_gs_slope_fut, tp_gs_slope_past)
#
tp_gs_slope_change <- tp_gs_slope_fut - tp_gs_slope_past
plot(tp_gs_slope_change)
#
writeRaster(tp_gs_slope_change, filename = 'data/tp_gs_slope_change.tif', overwrite=TRUE)

# answer Q4:
t2m_gs <- terra::rast('data/t2m_gs_2005_2023.tif')   
t2m_gs_slope_past <- terra::app(t2m_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
#
t2m_gs <- terra::rast('data/tas_gs_2026_2050.tif')   
t2m_gs_slope_fut <- terra::app(t2m_gs, function(x) {ifelse(any(is.na(x)), NA, sens.slope(x)$estimate / sd(x))})
#
t2m_gs_slope_fut <- terra::resample(t2m_gs_slope_fut, t2m_gs_slope_past)
#
t2m_gs_slope_change <- t2m_gs_slope_fut - t2m_gs_slope_past
plot(t2m_gs_slope_change)
#
writeRaster(t2m_gs_slope_change, filename = 'data/t2m_gs_slope_change.tif', overwrite=TRUE)


```
